backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
litellm==1.52.0
python-dotenv==1.0.1
pydantic==2.10.0
websockets==14.1
httpx==0.28.0
python-multipart==0.0.20

# Logging and monitoring
structlog==24.4.0

# Audio transcription
deepgram-sdk==3.8.0
aiofiles==24.1.0

# Database
sqlalchemy==2.0.36
